pyproject.toml
[project]
name = "langraph-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]
